started cleaning the DAS code and separated dasgenerator into two files to make it easier to maintain and change
